+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## 12.4.0-preview.1 (Unreleased)
+- Fixed bug where DataLakeFileClient.Upload() would deadlock if the content stream's position was not 0.
 - Fixed bug in DataLakeFileClient.OpenRead() causing us to do more download called than necessary.
 
 ## 12.3.1 (2020-08-18)